pyproject.toml
[tool.poetry]
name = "webscrap"
version = "0.1.0"
description = ""
authors = ["lksprado <lks-prado@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
lxml = "^5.1.0"
html5lib = "^1.1"
ipykernel = "^6.29.3"
selenium = "^4.18.1"
jupyter = "^1.0.0"
pandas = "^2.2.1"
pyinstaller = "^6.5.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
